fix mappings

+++ templates/noKibana/accountsESDTHistory.go
@@ -37,7 +37,7 @@ var AccountsESDTHistory = Object{
 				"type": "keyword",
 			},
 			"tokenNonce": Object{
-				"type": "long",
+				"type": "double",
 			},
 		},
 	},
